Moving api and services to inside src
